web/backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
drf-spectacular==0.27.0
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.0
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
